pyproject.toml
[project]
name = "diagram-generator"
version = "0.1.0"
description = "AWS Architecture Diagram Generator with LLM-powered agents"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Eden Djanashvili"}
]
dependencies = [
    "python-dotenv>=1.0.0",
]


[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
]

backend = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "langchain>=0.3.23",
    "langchain-openai>=0.3.12",
    "diagrams>=0.24.4",
    "pydantic>=2.11.2",
]

frontend = [
    "streamlit>=1.44.1",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"